requirements.txt
astroid==2.4.1
azure-cognitiveservices-vision-face==0.4.0
azure-common==1.1.25
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Flask==1.1.2
idna==2.9
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msrest==0.6.13
oauthlib==3.1.0
Pillow==7.1.2
pylint==2.5.2
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
